Refatorar nomenclatura utilizada no projeto
Evento -> Crise
Eventos -> Crises
novo evento -> nova crise
