+++ Alertas/alertaApp/src/main/res/layout/activity_cadastrar_crise.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="br.ita.bditac.mobile.alertas.CadastrarEventoActivity">
+    tools:context=".CadastrarCriseActivity">
 
     <LinearLayout
         android:layout_width ="match_parent"
