+++ Alertas/alertaApp/src/main/res/layout/activity_consulta_alertas.xml
@@ -3,7 +3,7 @@
             xmlns:tools="http://schemas.android.com/tools"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context="br.ita.bditac.mobile.alertas.ConsultaEventoActivity">
+            tools:context=".ConsultaAlertasActivity">
 
     <LinearLayout
         android:layout_width ="match_parent"
