+++ Alertas/alertaWSClient/src/test/java/br/ita/bditac/ws/client/alerta/AllTests.java
@@ -5,11 +5,9 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
-import br.ita.bditac.ws.model.Indicadores;
-
 
 @RunWith(Suite.class)
-@SuiteClasses({ EventoTests.class, AlertaTests.class, IndicadoresTests.class })
+@SuiteClasses({ CriseTests.class, AlertaTests.class, IndicadoresTests.class })
 public class AllTests {
 
 }
