+++ Alertas/alertaWSClient/src/test/java/br/ita/bditac/ws/client/alerta/CriseTests.java
@@ -7,12 +7,12 @@ import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
 
-import br.ita.bditac.ws.client.EventoClient;
-import br.ita.bditac.ws.model.Evento;
+import br.ita.bditac.ws.client.CriseClient;
+import br.ita.bditac.ws.model.Crise;
 
 
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class EventoTests extends TestCase {
+public class CriseTests extends TestCase {
 
     private static final String HOST_URL = "http://localhost:8080";
 
@@ -73,8 +73,8 @@ public class EventoTests extends TestCase {
      */
     @Test
     public void test01PostEvento() {
-        EventoClient eventoClient = new EventoClient(HOST_URL);
-        Evento eventoNovo = new Evento(
+        CriseClient criseClient= new CriseClient(HOST_URL);
+        Crise criseNova= new Crise(
                 "Deslizamento na favela do Paraiso",
                 0,
                 "Ze das Couves",
@@ -82,9 +82,9 @@ public class EventoTests extends TestCase {
                 "(12) 99876-1234",
                 -25.0D,
                 -45.0D);
-        Evento eventoRetorno = eventoClient.addEvento(eventoNovo);
-        assertNotNull(eventoRetorno);
-        assertEquals("Resposta(descricao):'" + eventoRetorno.getDescricao() + "' do POST diferente do que foi enviado: '" + eventoNovo.getDescricao() + "'!", eventoRetorno.getDescricao(), eventoNovo.getDescricao());
+        Crise criseRetorno= criseClient.addEvento(criseNova);
+        assertNotNull(criseRetorno);
+        assertEquals("Resposta(descricao):'" + criseRetorno.getDescricao() + "' do POST diferente do que foi enviado: '" + criseNova.getDescricao() + "'!", criseRetorno.getDescricao(), criseNova.getDescricao());
     }
 
     /**
@@ -127,10 +127,10 @@ public class EventoTests extends TestCase {
      */
     @Test
     public void test02GetEventoById() {
-        EventoClient eventoClient = new EventoClient(HOST_URL);
-        Evento evento = eventoClient.getEventoById(1);
-        assertNotNull(evento);
-        assertEquals("Resposta(descricao):'" + evento.getDescricao() + "' do POST diferente do que foi enviado!", evento.getDescricao(), "Deslizamento na favela do Paraiso");
+        CriseClient criseClient= new CriseClient(HOST_URL);
+        Crise crise= criseClient.getEventoById(1);
+        assertNotNull(crise);
+        assertEquals("Resposta(descricao):'" + crise.getDescricao() + "' do POST diferente do que foi enviado!", crise.getDescricao(), "Deslizamento na favela do Paraiso");
     }
 
     /**
@@ -156,9 +156,9 @@ public class EventoTests extends TestCase {
      */
     @Test
     public void test03GetEventoByIdInexistente() {
-        EventoClient eventoClient = new EventoClient(HOST_URL);
-        Evento evento = eventoClient.getEventoById(100);
-        assertNull(evento);
+        CriseClient criseClient= new CriseClient(HOST_URL);
+        Crise crise= criseClient.getEventoById(100);
+        assertNull(crise);
     }
     
 }
